Remove ITEMCOLLECTION from STACObjectTypes
